+++ pollux/lib/vc-jwt/src/main/scala/io/iohk/atala/pollux/vc/jwt/demos/JwtPresentationVerificationDemo.scala
@@ -260,6 +260,7 @@ object JwtPresentationVerificationDemo extends ZIOAppDefault {
           JwtPresentation.PresentationVerificationOptions(
             verifySignature = true,
             verifyDates = true,
+            verifyHoldersBinding = true,
             leeway = Duration.ZERO,
             Some(CredentialVerification.CredentialVerificationOptions(verifySignature = true, verifyDates = true))
           )